+++ src/MultipleSelect/action.js
@@ -7,6 +7,8 @@
  * of patent rights can be found in the PATENTS file in the same directory.
  */
 
+import warning from 'warning'
+
 export default {
 
   handleLabelRemove(value, e) {
@@ -25,6 +27,10 @@ export default {
   },
 
   handleLoad(data) {
+    if (this.props.dataFilter) {
+      data = this.props.dataFilter(data)
+      warning(!!data, '`dataFilter` should return new data, check the `dataFilter` of `Select`.')
+    }
     this.setState({ data })
   },